requirements.txt
Flask==2.3.3
Pillow==10.0.1
google-cloud-aiplatform==1.35.0
google-cloud-storage==2.10.0
Werkzeug==2.3.7
python-dotenv==1.0.0
